graph_rag/requirements.txt
python-dotenv
llama-index
llama-index-core
llama-index-llms-openai
llama-index-embeddings-openai
Flask
llama-index-vector-stores-qdrant
fastembed
